requirements.txt
streamlit
requests
beautifulsoup4
fpdf
transformers
scikit-learn
email_validator
python-dotenv
torch
accelerate